+++ spec/kafka/ffi/config_spec.rb
@@ -134,7 +134,7 @@ RSpec.describe Kafka::FFI::Config do
     callback = lambda {}
 
     expect { config.set_background_event_cb(&callback) }
-      .to change { config.get("background_event_cb") }.from(:unknown)
+      .to change { config.get("background_event_cb") }.from(:unknown).to(callback)
   ensure
     config.destroy
   end
@@ -144,7 +144,7 @@ RSpec.describe Kafka::FFI::Config do
     callback = lambda {}
 
     expect { config.set_dr_msg_cb(&callback) }
-      .to change { config.get("dr_msg_cb") }.from(:unknown)
+      .to change { config.get("dr_msg_cb") }.from(:unknown).to(callback)
   ensure
     config.destroy
   end
@@ -154,7 +154,7 @@ RSpec.describe Kafka::FFI::Config do
     callback = lambda {}
 
     expect { config.set_consume_cb(&callback) }
-      .to change { config.get("consume_cb") }.from(:unknown)
+      .to change { config.get("consume_cb") }.from(:unknown).to(callback)
   ensure
     config.destroy
   end
@@ -164,7 +164,7 @@ RSpec.describe Kafka::FFI::Config do
     callback = lambda {}
 
     expect { config.set_rebalance_cb(&callback) }
-      .to change { config.get("rebalance_cb") }.from(:unknown)
+      .to change { config.get("rebalance_cb") }.from(:unknown).to(callback)
   ensure
     config.destroy
   end
@@ -174,7 +174,7 @@ RSpec.describe Kafka::FFI::Config do
     callback = lambda {}
 
     expect { config.set_offset_commit_cb(&callback) }
-      .to change { config.get("offset_commit_cb") }.from(:unknown)
+      .to change { config.get("offset_commit_cb") }.from(:unknown).to(callback)
   ensure
     config.destroy
   end
@@ -184,7 +184,7 @@ RSpec.describe Kafka::FFI::Config do
     callback = lambda {}
 
     expect { config.set_error_cb(&callback) }
-      .to change { config.get("error_cb") }.from(:unknown)
+      .to change { config.get("error_cb") }.from(:unknown).to(callback)
   ensure
     config.destroy
   end
@@ -194,7 +194,7 @@ RSpec.describe Kafka::FFI::Config do
     callback = lambda {}
 
     expect { config.set_throttle_cb(&callback) }
-      .to change { config.get("throttle_cb") }.from(:unknown)
+      .to change { config.get("throttle_cb") }.from(:unknown).to(callback)
   ensure
     config.destroy
   end
@@ -204,7 +204,7 @@ RSpec.describe Kafka::FFI::Config do
     callback = lambda {}
 
     expect { config.set_log_cb(&callback) }
-      .to change { config.get("log_cb") }
+      .to change { config.get("log_cb") }.from(:unknown).to(callback)
   ensure
     config.destroy
   end
@@ -214,7 +214,7 @@ RSpec.describe Kafka::FFI::Config do
     callback = lambda {}
 
     expect { config.set_stats_cb(&callback) }
-      .to change { config.get("stats_cb") }.from(:unknown)
+      .to change { config.get("stats_cb") }.from(:unknown).to(callback)
   ensure
     config.destroy
   end
@@ -224,7 +224,7 @@ RSpec.describe Kafka::FFI::Config do
     callback = lambda {}
 
     expect { config.set_oauthbearer_token_refresh_cb(&callback) }
-      .to change { config.get("oauthbearer_token_refresh_cb") }.from(:unknown)
+      .to change { config.get("oauthbearer_token_refresh_cb") }.from(:unknown).to(callback)
   ensure
     config.destroy
   end
@@ -234,7 +234,7 @@ RSpec.describe Kafka::FFI::Config do
     callback = lambda {}
 
     expect { config.set_socket_cb(&callback) }
-      .to change { config.get("socket_cb") }
+      .to change { config.get("socket_cb") }.from(:unknown).to(callback)
   ensure
     config.destroy
   end
@@ -244,7 +244,7 @@ RSpec.describe Kafka::FFI::Config do
     callback = lambda {}
 
     expect { config.set_connect_cb(&callback) }
-      .to change { config.get("connect_cb") }.from(:unknown)
+      .to change { config.get("connect_cb") }.from(:unknown).to(callback)
   ensure
     config.destroy
   end
@@ -254,7 +254,7 @@ RSpec.describe Kafka::FFI::Config do
     callback = lambda {}
 
     expect { config.set_closesocket_cb(&callback) }
-      .to change { config.get("closesocket_cb") }.from(:unknown)
+      .to change { config.get("closesocket_cb") }.from(:unknown).to(callback)
   ensure
     config.destroy
   end
@@ -264,7 +264,7 @@ RSpec.describe Kafka::FFI::Config do
     callback = lambda {}
 
     expect { config.set_open_cb(&callback) }
-      .to change { config.get("open_cb") }
+      .to change { config.get("open_cb") }.from(:unknown).to(callback)
   ensure
     config.destroy
   end
@@ -274,7 +274,7 @@ RSpec.describe Kafka::FFI::Config do
     callback = lambda {}
 
     expect { config.set_ssl_cert_verify_cb(&callback) }
-      .to change { config.get("ssl.certificate.verify_cb") }.from(:unknown)
+      .to change { config.get("ssl.certificate.verify_cb") }.from(:unknown).to(callback)
   ensure
     config.destroy
   end